refactor(context_compressor): extract token-budget cluster (LB3)
Window 420-649 (current numbering; pin 616-868 minus the text-util leaf)
into context_compressor_budget.py. HISTORICAL_TASK_HEADING moves with the
cluster (sole external dep, re-exported is-identical). 29 members
re-exported through the godfile. 39/39 targeted tests green.

Part of #78645
Part of #78647
